requirements.txt
tweepy
requests
textblob
python-dotenv